Add numpy as a build requirement
To prevent an incompatibility with the changes to numpy integers in
numpy 2.0 (i.e., `cnp.int_t` does not exist anymore; we could just use
`int` there, as far as I understand), we must pin the numpy version for
the build requirements to 1.26. (For now.)

The install_requirements (such as MDAnalysis) lead to the numpy install
requirement being interpreted as ~1.26, so on the runtime end things are
fine. The problem may have been screened away by build tests on systems
that have a system-wide numpy<2.0 installation.